// kysely/src/types/database.d.ts
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */
import type { ColumnType } from 'kysely'

export type Generated<T> =
  T extends ColumnType<infer S, infer I, infer U>
    ? ColumnType<S, I | undefined, U>
    : ColumnType<T, T | undefined, T>

export type Timestamp = ColumnType<Date, Date | string, Date | string>

export interface Todos {
  completed: Generated<boolean>
  createdAt: Generated<Timestamp>
  id: Generated<number>
  name: string
  userId: number
}

export interface Users {
  createdAt: Generated<Timestamp>
  email: string
  id: Generated<number>
  name: string
  password: string
}

export interface DB {
  todos: Todos
  users: Users
}